added register page with authlayout
