--- src/app/login/page.jsx
@@ -1,36 +1,0 @@
-import { GalleryVerticalEnd } from "lucide-react"
-
-import { LoginForm } from "@/components/login-form";
-import Link from "next/link";
-import LoginLottie from "./LoginLottie";
-
-const LoginPage = () => {
-
-  return (
-    <div className="grid min-h-svh md:grid-cols-2">
-      <div className="flex flex-col gap-4 p-6 md:p-10 bg-muted">
-        <div className="flex justify-center gap-2 md:justify-start">
-          <Link href="/" className="flex items-center gap-2 font-medium">
-            <div className="bg-primary text-primary-foreground flex size-6 items-center justify-center rounded-md">
-              <GalleryVerticalEnd className="size-4" />
-            </div>
-            
-            <span className='text-xl font-bold'>QuizPilot</span>
-          </Link>
-        </div>
-
-        <div className="mt-10 md:mt-0 flex flex-1 md:items-center justify-center">
-          <div className="w-full max-w-xs">
-            <LoginForm />
-          </div>
-        </div>
-
-      </div>
-      <div className="bg-background hidden md:flex md:items-center md:justify-center">
-        <LoginLottie/>
-      </div>
-    </div>
-  );
-};
-
-export default LoginPage;
