+++ src/components/google/Google.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Button } from '../ui/button'
 import { signIn } from 'next-auth/react'
 
-export default function Google() {
+export default function Google({method}) {
     const handleGoogleLogin = () => {
         signIn("google",{callbackUrl : "/"});
     }
@@ -22,7 +22,7 @@ export default function Google() {
                     <path fill="#4A90E2" d="M124.4 329.2c-10.4-30.9-10.4-64.1 0-95l.1-68.5H32.8c-42.7 85.4-42.7 184.7 0 270.1l91.6-68.6z" />
                     <path fill="#FBBC05" d="M272 106.3c39.4-.6 77.3 14 106.2 41.4l79.3-79.3C412.7 23 344.6-1.3 272 0 169.3 0 77.5 51.3 32.8 146.3l91.6 68.5C145.2 152.7 203.4 106.3 272 106.3z" />
                 </svg>
-                Login with Google
+                {method} with Google
             </Button>
         </>
     )
